+++ lib/generators/thumbs_up/templates/migration.rb
@@ -1,7 +1,7 @@
 class ThumbsUpMigration < ActiveRecord::Migration
   def self.up
     create_table :votes, :force => true do |t|
-      t.boolean    :vote, :default => false
+      t.integer    :vote, :null => false
       t.references :voteable, :polymorphic => true, :null => false
       t.references :voter,    :polymorphic => true
       t.timestamps